Refactoring Widget inicial
